.gemini/antigravity/brain/c718aa18-28de-4cd5-9403-f4be6c1ae8db/Job_Search_Agent_System/requirements.txt
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
openai>=1.0.0
groq>=0.4.0
numpy>=1.24.0
playwright>=1.40.0
python-telegram-bot>=20.7
